Fixed bug: when failed to submit a new job, the submit result is empty, so attempting to extract the jobID will fail. In this case, do not attempt to delete the job, since there is no job ID. ******************************************************************8 Job.m: updated jobID() to return -1 for failure. JobManager.m: updated deleteJob to check return value from jobID(). updated log print ni startJob()
git-svn-id: https://matano1-tp.haifa.ibm.com/svn/taco@582 4c94cbd0-a001-4ae1-9522-79914504ef8a

+++ Experiments/Jobs/JobManager.m
@@ -303,13 +303,18 @@ methods (Static)
         job.submitResult = result;
         job.lastLogFileSize = 0;
         job.idleCount = 0;
-        Logger.log(['JobManager::startJob. result = ' result]);
+        Logger.log(['JobManager::startJob. result = ''' result '''']);
     end
     
     %% deleteJob
     
     function deleteJob(job)
         id = job.jobID();
+        if (id == -1)
+            Logger.log(['Error deleting job run. job id =  ' ...
+                         num2str(id)]);
+            return;
+        end
         deleteCommand = ['qdel ' num2str(id) ];
         Logger.log(['deleteCommand = "' deleteCommand '"']);
         [status, result] = system(deleteCommand);
